refactor(cli): extract execSwingsetTransaction
